requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
djangorestframework==3.12.4
kombu==5.1.0
Pillow==8.2.0
prompt-toolkit==3.0.19
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
